devex: remove deprecated env-destroy-utils
