--- env-destroy-utils/sleep.ts
@@ -1,3 +1,0 @@
-export const sleepForMilliseconds = (ms: number): Promise<void> => {
-  return new Promise(resolve => setTimeout(resolve, ms));
-};
